backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
aiofiles==23.2.1
python-dotenv==1.0.1
loguru==0.7.2

# Image / PDF
pymupdf==1.24.3
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-image==0.23.2
scipy==1.13.0

# AI / ML (CPU — replace +cpu with +cu118 for GPU)
torch==2.3.0+cpu
torchvision==0.18.0+cpu
transformers==4.41.0
datasets==2.19.1
accelerate==0.30.0
sentencepiece==0.2.0
protobuf==3.20.3
evaluate==0.4.2

# Exports
reportlab==4.2.0
python-docx==1.1.2

# Storage
boto3==1.34.111
minio==7.2.7

# Test
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
